requirements.txt
africastalking==1.2.4
certifi==2021.5.30
charset-normalizer==2.0.4
contextlib2==21.6.0
idna==3.2
kafka-python==2.0.2
python-dotenv==0.19.0
requests==2.26.0
schema==0.7.4
urllib3==1.26.6
